+++ packages/orm/angel_migration/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log
 
-## 8.2.2
+## 8.3.0
 
-* Added `MutableTable`
+* Added `addIndex`, `dropIndex` and `dropPrimaryIndex`
 
 ## 8.2.1