+++ packages/orm/angel_migration_runner/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log
 
-## 8.2.3
+## 8.3.0
 
-* Added index support
+* Added `addIndex`, `dropIndex` and `dropPrimaryIndex`
 
 ## 8.2.2